[DEBT-043-D10] Split exa/websets/client.py into focused modules
- Split exa/websets/client.py (442 LoC) into mixins:
  - _http.py (219): HTTP client lifecycle, request/retry, Retry-After parsing
  - _websets.py (117): Webset CRUD operations
  - _items.py (80): Webset item operations
  - _searches.py (100): Webset search operations
  - client.py (54): Thin facade composing mixins
- Used mixin composition via multiple inheritance
- All files now ≤400 lines
- Quality gates passed: pre-commit, mypy, pytest (1052 tests)
